+++ gui/tasks/distribution.js
@@ -138,17 +138,18 @@ const config = {
     artifactName: 'MullvadVPN-${version}_${arch}.${ext}',
     publisherName: 'Mullvad VPN AB',
     extraResources: [
-      { from: distAssets('mullvad.exe'), to: '.' },
-      { from: distAssets('mullvad-problem-report.exe'), to: '.' },
-      { from: distAssets('mullvad-daemon.exe'), to: '.' },
-      { from: distAssets('talpid_openvpn_plugin.dll'), to: '.' },
+      { from: distAssets(path.join(getWindowsDistSubdir(), 'mullvad.exe')), to: '.' },
+      { from: distAssets(path.join(getWindowsDistSubdir(), 'mullvad-problem-report.exe')), to: '.' },
+      { from: distAssets(path.join(getWindowsDistSubdir(), 'mullvad-daemon.exe')), to: '.' },
+      { from: distAssets(path.join(getWindowsDistSubdir(), 'talpid_openvpn_plugin.dll')), to: '.' },
       {
-        from: root(path.join('windows', 'winfw', 'bin', '${env.CPP_BUILD_TARGET}-${env.CPP_BUILD_MODE}', 'winfw.dll')),
+        from: root(path.join('windows', 'winfw', 'bin', getWindowsTargetArch() + '-${env.CPP_BUILD_MODE}', 'winfw.dll')),
         to: '.',
       },
-      { from: distAssets(path.join('binaries', getWindowsTargetSubdir(), 'openvpn.exe')), to: '.' },
-      { from: distAssets(path.join('binaries', getWindowsTargetSubdir(), 'apisocks5.exe')), to: '.' },
-      { from: distAssets(path.join('binaries', getWindowsTargetSubdir(), 'wintun/wintun.dll')), to: '.' },
+      // OpenVPN, APISocks5 and Wintun do not have ARM64 builds yet.
+      { from: distAssets(path.join('binaries/x86_64-pc-windows-msvc/openvpn.exe')), to: '.' },
+      { from: distAssets(path.join('binaries/x86_64-pc-windows-msvc/apisocks5.exe')), to: '.' },
+      { from: distAssets(path.join('binaries/x86_64-pc-windows-msvc/wintun/wintun.dll')), to: '.' },
       {
         from: distAssets(path.join('binaries', getWindowsTargetSubdir(), 'split-tunnel/mullvad-split-tunnel.sys')),
         to: '.'
@@ -254,6 +255,18 @@ function packWin() {
       beforeBuild: (options) => {
         process.env.CPP_BUILD_MODE = release ? 'Release' : 'Debug';
         process.env.CPP_BUILD_TARGET = options.arch;
+        switch (options.arch) {
+          case 'x64':
+            process.env.TARGET_TRIPLE = 'x86_64-pc-windows-msvc';
+            process.env.SETUP_SUBDIR = '.';
+            break;
+          case 'arm64':
+            process.env.TARGET_TRIPLE = 'aarch64-pc-windows-msvc';
+            process.env.SETUP_SUBDIR = 'aarch64-pc-windows-msvc';
+            break;
+          default:
+            throw new Error(`Invalid or unknown target (only one may be specified)`);
+        }
         return true;
       },
       afterAllArtifactBuild: (buildResult) => {
@@ -396,6 +409,14 @@ function root(relativePath) {
   return path.join(path.resolve(__dirname, '../../'), relativePath);
 }
 
+function getWindowsDistSubdir() {
+  if (targets === 'aarch64-pc-windows-msvc') {
+    return targets;
+  } else {
+    return '';
+  }
+}
+
 function getWindowsTargetArch() {
   if (targets && process.platform === 'win32') {
     if (targets === 'aarch64-pc-windows-msvc') {
@@ -403,7 +424,7 @@ function getWindowsTargetArch() {
     }
     throw new Error(`Invalid or unknown target (only one may be specified)`);
   }
-  // Use host architecture.
+  // Use host architecture (we assume this is x64 since building on Arm64 isn't supported).
   return 'x64';
 }
 
@@ -414,7 +435,7 @@ function getWindowsTargetSubdir() {
     }
     throw new Error(`Invalid or unknown target (only one may be specified)`);
   }
-  // Use host architecture.
+  // Use host architecture (we assume this is x64 since building on Arm64 isn't supported).
   return 'x86_64-pc-windows-msvc';
 }